validate uniqueness of entries
this should validate that no two entries are the same, it compares the
run number with the operator id.  I made a couple of assumptions here.
1 - that all run numbers are unique to a specific line, and there won’t
ever be a common run number b/w different trains.  For example an El
Train and a Metra could never share a run number
2 - I’m also assuming route names have unique run numbers so a Brown
and purple line on the El couldn’t share a number
3 - This would display the same train run numbers with different
operators, I believe this means that the runs are on different dates or
times.  As there isn’t any input for that, its def something I would
talk about adding in the future.