add import method to train model
this model takes the csv file and goes through row by row and takes the
values and stores them in a temporary hash, lowercase_hash.  The
purpose of the lowercase_hash is to take the headers and make them case
insensitive. For input into the database.
 I also had a problem with the route_name column header in the test
case provided.  I had extra whitespace at the beginning of the string,
so I found a ruby method, strip, that removes space characters from the
beginning and end of a string, I considered using a regular expression
for white space throughout a header string, but I thought this looks a
lot cleaner and it’s easier to read, it might be better to throw and
error and tell a user to input a cvs file with properly formatted
headers.
Because of the way I set up the db it throws a validation error when
duplicate data is input, so my next step will be to handle that error
and continue adding data